requirements.txt
# ── Núcleo numérico / utilidades ─────────────────────────────────────────────
numpy>=2.1,<3
pandas==2.3.0
pyarrow==20.0.0
orjson>=3.11.0,<4
ormsgpack>=1.10.0,<2
jiter>=0.4.0,<1
pydantic==2.11.7
python-dotenv==1.1.0
SQLAlchemy==2.0.43
requests==2.32.4
cachetools>=5.3,<6
unidecode>=1.3,<2

# ── I/O, documentos y medios ────────────────────────────────────────────────
pillow>=9.2,<12
pypdf==5.7.0
pymupdf==1.25.5
openpyxl==3.1.5
xlrd>=2.0.1,<3
docxtpl==0.20.0
unstructured[docx]==0.18.2
moviepy==2.2.1
proglog>=0.1.10,<1.1
jinja2==3.1.6
json-repair==0.41.0
pubchempy==1.0.4
lxml>=4.9,<6
pdfminer.six>=20231228

# ── Visualización ───────────────────────────────────────────────────────────
matplotlib==3.10.3
wordcloud==1.9.4
squarify==0.4.4
streamlit==1.45.1

# ── SDKs de modelos ─────────────────────────────────────────────────────────
openai>=1.60.0,<2
mistralai>=1.5.0,<2
sentence-transformers>=2.7.0

# ── LangChain / LangGraph (conjuntos compatibles) ───────────────────────────
langchain==0.3.27
langchain-core==0.3.74
langchain-community==0.3.27
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.3.45
langchain-experimental>=0.3.0,<0.4

langgraph==0.6.5
langgraph-supervisor==0.0.29
# Servidor local para `langgraph dev` (útil en desarrollo local)
langgraph-api==0.2.130

# ── Otros ───────────────────────────────────────────────────────────────────
trustcall
pytest
pytest-cov
setuptools>=70
